Major improvements in the projects solved some mejor flows and bugs

+++ src/main/java/com/blood/bank/Blood/bank/Repository/BloodCampRepository.java
@@ -1,10 +1,9 @@
 package com.blood.bank.Blood.bank.Repository;
 
+import com.blood.bank.Blood.bank.model.BloodCamp;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.blood.bank.Blood.bank.model.BloodCamp;
-
 @Repository
-public interface BloodCampRepository extends JpaRepository<BloodCamp, Integer> {}
-
+public interface BloodCampRepository extends JpaRepository<BloodCamp, Long> {
+}
